Moved three more classes
